// LDC.Atlas.WebApp/Client/app/shared/entities/counterparty-bankaccounts.entity.ts
export class CounterpartyBankAccounts {
    randomId: number;
    bankAccountId: number;
    bankKey: string;
    bankName: string;
    bankCountryKey: number;
    bankCity: string;
    bankSwiftCode: string;
    nCC: string;
    nCS: string;
    bankBranch: string;
    bankZIPCode: string;
    bankAccountDesc: string;
    accountNo: string;
    accountCCY: string;
    externalReference: string;
    bankTypeID: number;
    bankAccountStatusID: number;
    isBankAccountDefault: boolean;
    isDeactivated: boolean;
}